decimal

+++ CurrencyApi/PublicApi/Helpers/CurrencyHelper.cs
@@ -4,5 +4,5 @@ namespace Fuse8.BackendInternship.PublicApi.Helpers;
 
 public static class CurrencyHelper
 {
-    public static decimal RoundCurrencyValue(double value, int accuracy) => Math.Round((decimal)value, accuracy);
+    public static decimal RoundCurrencyValue(decimal value, int accuracy) => Math.Round(value, accuracy);
 }
